Move callata logic to hooks module
